# fitness-tracker-backend/ft-api/src/main/resources/openapi/fitness-tracker-api.yaml
openapi: 3.0.3
info:
  title: fitness-tracker api
  description: fitness-tracker api
  version: 1.0.0
servers:
  - url: http://localhost:8081/fitnesstracker/api/v1
paths:
...
  /workout/save:
    summary: Endpoint for locally saving a workout while doing one
    post:
      operationId: savingWorkout
      requestBody:
        description: The workouttemplate to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWorkoutDto'
      responses:
        201:
          description: The workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewWorkoutDto'
...
components:
  schemas:
    NewWorkoutDto:
      type: object
      properties:
        name:
          type: string
        setInfoList:
          type: array
          items:
            $ref: '#/components/schemas/SetInfoDto'
...
    SetInfoDto:
      type: object
      properties:
        movementType:
          type: string
        amountOfTimes:
          type: integer
        orderOfSet:
          type: integer
        movement:
          $ref: '#/components/schemas/MovementDto'
        reps:
          type: array
          items:
            $ref: '#/components/schemas/RepInfoDto'
    RepInfoDto:
      type: object
      properties:
        reps:
          type: integer
        weight:
          type: number
        timestamp:
          type: string
...
    MovementDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        muscleGroup:
          type: string